Private API

Common methods

GridVisualize.SubVisualizer — Type

A SubVisualizer is just a dictionary which contains plotting information, including type of the plotter and its position in the plot.

GridVisualize.default_backend — Constant

nothing as initial default plotter

GridVisualize.MeshCatType — Type

abstract type MeshCatType

Abstract type for dispatching on plotter

Base.getindex — Method

getindex(p, i, j)

Return a SubVisualizer

Base.size — Method

size(p)

Return the layout of a GridVisualizer

GridVisualize.ismeshcat — Method

ismeshcat(Plotter)

Heuristically check if Plotter is MeshCat

GridVisualize.bregion_cmap — Method

bregion_cmap(n)

Create customized distinguishable colormap for boundary regions

GridVisualize.extract_visible_cells3D — Method

extract_visible_cells3D(grid, xyzcut; primepoints, Tp, Tf)

Extract visible tetrahedra - those intersecting with the planes x=xyzcut[1] or y=xyzcut[2] or z=xyzcut[3].

Return corresponding points and facets for each region for drawing as mesh (Makie,MeshCat) or trisurf (pyplot)

GridVisualize.marching_tetrahedra — Method

marching_tetrahedra(coord, cellnodes, func, planes, flevels; tol, primepoints, primevalues, Tv, Tp, Tf)

GridVisualize.marching_tetrahedra — Method

marching_tetrahedra(grid, func, planes, flevels; kwargs...)

GridVisualize.markerpoints — Method

markerpoints(points, nmarkers, transform)

Assume that points are nodes of a polyline. Place nmarkers equidistant markers at the polyline, under the assumption that the points are transformed via the transformation matrix M vor visualization.

GridVisualize.region_cmap — Method

region_cmap(n)

Create customized distinguishable colormap for interior regions

GridVisualize.tet_x_plane! — Method

tet_x_plane!(ixcoord, ixvalues, pointlist, node_indices, planeq_values, function_values; tol)

Calculate intersections between tetrahedron with given linear fucntion data and plane

Adapted from https://github.com/j-fu/gltools/blob/master/glm-3d.c#L341

A non-empty intersection is either a triangle or a planar quadrilateral, define by either 3 or 4 intersection points between tetrahedron edges and the plane.

Input:

  pointlist: 3xN array of grid point coordinates

nodeindices: 4 element array of node indices (pointing into pointlist and functionvalues)

planeq_values: 4 element array of plane equation evaluated at the node coordinates

function_values: N element array of function values

Mutates:

ixcoord: 3x4 array of plane - tetedge intersection coordinates

ixvalues: 4 element array of fuction values at plane - tetdedge intersections

Returns:

nxs,ixcoord,ixvalues

This method can be used both for the evaluation of plane sections and for the evaluation of function isosurfaces.

PyPlot

GridVisualize.tridata — Method

tridata(grid)

Return tridata to be splatted to PyPlot calls

Makie

GridVisualize.makeaxis3d — Method

makeaxis3d(ctx)

Dispatch between LScene and new Axis3. Axis3 does not allow zoom, so we support LScene in addition.

GridVisualize.makescene2d — Method

makescene2d(ctx)

Complete scene with title and status line showing interaction state. This uses a gridlayout and its protrusion capabilities.

GridVisualize.scene_interaction — Function

 scene_interaction(update_scene,view,switchkeys::Vector{Symbol}=[:nothing])

Control multiple scene elements via keyboard up/down keys. Each switchkey is assumed to correspond to one of these elements. Pressing a switch key transfers control to its associated element.

Control of values of the current associated element is performed by triggering change values via up/down (± 1) resp. pageup/pagedown (±10) keys

The update_scene callback gets passed the change value and the symbol.

GridVisualize.FlippableLayout — Module

This module is a submodule of GridVisualize.

It manages a layoutscene with interactive layout and blocking functionality.

Thanks to Julius Krumbiegel for providing a basic implementation of focus switching.

GridVisualize avoids creating dependencies on plotting backends. So we provide a way to emulate "import Makie" by allowing to set it as a global variable in the setmakie!. As a consequence, we can't use Makie types at compile time.

GridVisualize.FlippableLayout.FLayout — Type

mutable struct FLayout

Struct describing flippable layout data. We don't type annotate with Makie types as they are unknown at start time.

visible

Visible GridLayout

offscreen

Hidden GridLayot

blocked

Has the layout been blocked by the block key ?

layoutables

Layoutables attached to layout

condition

Condition variable working together with the blocked field.

GridVisualize.FlippableLayout.flayoutscene — Method

flayoutscene(;blocked=false, kwargs...)

Layoutscene with interactive layout and blocking functionality.

The , key switches between focused view showing only one subscene and "gallery view" showing all layoutables at once.

The space key toggles blocking of the execution of the main therad when yield is replaced by yieldwait. Initial blocking state is set by the blocked kwarg.

The kwargs... are the same as of AbstractPlotting.layoutscene.

The idea is that this can work in some cases as a drop-in replacement of layoutscene.

GridVisualize.FlippableLayout.setmakie! — Method

setmakie!(MyMakie)

Set the Makie module. This Makie can be GLMakie,WGLMakie,CairoMakie

GridVisualize.FlippableLayout.yieldwait — Method

 yieldwait(fliplayoutscene)

Yield and wait in case of scene being blocked via space key toggle

Plots

GridVisualize.rectdata — Method

rectdata(grid, U)

Return rectangular grid data + function to be splatted into Plots calls

VTKView