Public API

GridVisualize.available_kwargs — Method

available_kwargs()

Available kwargs for all methods of this package.

show: Show plot immediately. Default: false

reveal: Show plot immediately (same as :show). Default: false

clear: Clear plot before new plot.. Default: true

layout: Layout of plots in window. Default: (1, 1)

resolution: Plot window resolution. Default: (500, 500)

legend: Legend (position): one of [:none, :best, :lt, :ct, :rt, :lc, :rc, :lb, :cb, :rb]. Default: none

title: Plot title. Default: ``

xlabel: x axis label. Default: ``

ylabel: y axis label. Default: ``

zlabel: z axis label. Default: ``

xlimits: x limits. Default: (1, -1)

ylimits: y limits. Default: (1, -1)

zlimits: z limits. Default: (1, -1)

flimits: function limits. Default: (1, -1)

aspect: Aspect ratio modification. Default: 1.0

fontsize: Fontsize of titles. All others are relative to it. Default: 20

linewidth: 1D plot or isoline linewidth. Default: 2

linestyle: 1D Plot linestyle: one of [:solid, :dash, :dot, :dashdot, :dashdotdot]. Default: solid

markevery: 1D plot marker stride. Default: 5

markersize: 1D plot marker size. Default: 5

markershape: 1D plot marker shape: one of [:none, :circle, :star5, :diamond, :hexagon, :cross, :xcross, :utriangle, :dtriangle, :rtriangle, :ltriangle, :pentagon, :+, :x]. Default: none

color: 1D plot line color. Default: (0.0, 0.0, 0.0)

cellwise: 1D plots cellwise can be slow). Default: false

label: 1D plot label. Default: ``

isolines: 2D contour plot: number of isolines. Default: 11

elevation: 2D plot height factor for elevation. Default: 0.0

colorlevels: 2D/3D contour plot: number of color levels. Default: 51

colormap: 2D/3D contour plot color map (any from ColorSchemes.jl). Default: viridis

colorbar: 2D/3D plot colorbar. One of [:none, :vertical, :horizontal]. Default: vertical

alpha: 3D outline surface alpha value. Default: 0.1

interior: 3D plot interior of grid. Default: true

outline: 3D plot outline of domain. Default: true

xplane: 3D x plane position. Default: 1.7976931348623157e308

yplane: 3D y plane position. Default: 1.7976931348623157e308

zplane: 3D z plane position. Default: 1.7976931348623157e308

flevel: 3D isosurface level. Default: 1.7976931348623157e308

azim: 3D azimuth angle (in degrees). Default: -60

elev: 3D elevation angle (in degrees). Default: 30

perspectiveness: 3D perspective A number between 0 and 1, where 0 is orthographic, and 1 full perspective. Default: 0.25

scene3d: 3D plot type of Makie scene. Alternaitve to Axis3 is LScene. Default: Axis3

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

subplot: Private: Actual subplot. Default: (1, 1)

backend: Backend for PlutoVista plot. Default: default

GridVisualize.default_plotter! — Method

   plotter!(Plotter)

Set plotter module as the default plotter backend.

GridVisualize.default_plotter — Method

default_plotter()

Return default plotter backend

GridVisualize.gridplot! — Method

gridplot!(visualizer[i,j], grid, kwargs...)
gridplot!(visualizer, grid, kwargs...)

Plot grid into subplot in the visualizer. If [i,j] is omitted, [1,1] is assumed.

Keyword arguments: see available_kwargs

GridVisualize.gridplot — Method

gridplot(grid; Plotter=default_plotter(); kwargs...)

Create grid visualizer and plot grid

Keyword arguments: see available_kwargs

GridVisualize.ismakie — Method

ismakie(Plotter)

Heuristically check if Plotter is Makie/WGLMakie

GridVisualize.isplots — Method

isplots(Plotter)

Heuristically check if Plotter is Plots

GridVisualize.isplutovista — Method

isplutovista(Plotter)

Heuristically check if Plotter is MeshCat

GridVisualize.ispyplot — Method

ispyplot(Plotter)

Heuristically check if Plotter is PyPlot

GridVisualize.isvtkview — Method

isvtkview(Plotter)

Heuristically check if Plotter is VTKView

GridVisualize.plottertype — Method

plottertype(p)

Return the type of a plotter.

GridVisualize.plottertype — Method

plottertype(Plotter)

Heuristically detect type of plotter, returns the corresponding abstract type fro plotting.

GridVisualize.reveal — Method

reveal(visualizer)

Finish and show plot. Same as setting :reveal=true or :show=true in last plot statment for a context.

GridVisualize.save — Method

save(fname, scene; Plotter)

Save scene returned from reveal, scalarplot or gridplot to disk.

GridVisualize.save — Method

save(fname, visualizer)

Save last plotted figure from visualizer to disk.

GridVisualize.scalarplot! — Method

scalarplot!(visualizer[i,j], grid, vector; kwargs...)
scalarplot!(visualizer, grid, vector; kwargs...)
scalarplot!(visualizer[i,j], grid, function; kwargs...)
scalarplot!(visualizer[i,j], X, vector; kwargs...)
scalarplot!(visualizer[i,j], X, function; kwargs...)
scalarplot!(visualizer[i,j], X, Y, function; kwargs...)
scalarplot!(visualizer[i,j], X, Y, Z, function; kwargs...)

Plot node vector on grid as P1 FEM function on the triangulation into subplot in the visualizer. If [i,j] is omitted, [1,1] is assumed.

If instead of the node vector, a function is given, it will be evaluated on the grid.

If instead of the grid, coordinate vectors are given, a temporary grid is created.

Keyword arguments: see available_kwargs

GridVisualize.scalarplot — Method

scalarplot(grid,vector; Plotter=default_plotter())
scalarplot(grid,function; Plotter=default_plotter())
scalarplot(X,vector; Plotter=default_plotter())
scalarplot(X,function; Plotter=default_plotter())
scalarplot(X,Y,function; Plotter=default_plotter())
scalarplot(X,Y,Z,function; Plotter=default_plotter())

Plot node vector on grid as P1 FEM function on the triangulation.

If instead of the node vector, a function is given, it will be evaluated on the grid.

If instead of the grid, vectors for coordinates are given, a grid is created automatically.

Keyword arguments: see available_kwargs

GridVisualize.GridVisualizer — Type

struct GridVisualizer

GridVisualizer struct

GridVisualize.GridVisualizer — Method

    GridVisualizer(; Plotter=default_plotter() , kwargs...)

Create a grid visualizer

Plotter: defaults to default_plotter() and can be PyPlot, Plots, VTKView, Makie. This pattern to pass the backend as a module to a plot function allows to circumvent to create heavy default package dependencies.

Depending on the layout keyword argument, a 2D grid of subplots is created. Further ...plot! commands then plot into one of these subplots:

vis=GridVisualizer(Plotter=PyPlot, layout=(2,2)
...plot!(vis[1,2], ...)

A ...plot command just implicitely creates a plot context:

gridplot(grid, Plotter=PyPlot) 

is equivalent to

vis=GridVisualizer(Plotter=PyPlot, layout=(1,1))
gridplot!(vis,grid) 
reveal(vis)

Please note that the return values of all plot commands are specific to the Plotter.

An interactive mode switch key for GLMakie (,) and VTKView (*) allows to toggle between "gallery view" showing all plots at once and "focused view" showing only one plot.

Keyword arguments: see available_kwargs

GridVisualize.MakieType — Type

abstract type MakieType

Abstract type for dispatching on plotter

GridVisualize.PlotsType — Type

abstract type PlotsType

Abstract type for dispatching on plotter

GridVisualize.PlutoVistaType — Type

abstract type PlutoVistaType

Abstract type for dispatching on plotter

GridVisualize.PyPlotType — Type

abstract type PyPlotType

Abstract type for dispatching on plotter

GridVisualize.VTKViewType — Type

abstract type VTKViewType

Abstract type for dispatching on plotter