Private API

Common methods

GridVisualize.AbstractPlotterType — Type

abstract type AbstractPlotterType

Abstract type for dispatching on plotter

Base.getindex — Method

getindex(p, i, j)

Return a SubVisualizer

Base.size — Method

size(p)

Return the layout of a GridVisualizer

GridVisualize.ismeshcat — Method

ismeshcat(Plotter)

Heuristically check if Plotter is MeshCat

GridVisualize.quiverdata — Method

quiverdata(
    rastercoord,
    rasterflux;
    vscale,
    vnormalize
) -> Tuple{Union{Base.ReshapedArray{Float32, 2}, Matrix{Float32}}, Union{Base.ReshapedArray{Float32, 2}, Matrix{Float32}}}

Extract nonzero fluxes for quiver plots from rastergrid.

Returns qc, qv - d x nquiver matrices.

If vnormalize is true, the vector field is normalized to vscale*min(spacing), otherwise, it is scaled by vscale Result data are meant to be ready for being passed to calls to quiver.

GridVisualize.vectorsample — Method

vectorsample(
    grid::ExtendableGrid{Tv, Ti},
    v;
    offset,
    spacing,
    reltol
) -> Tuple{Vector{Vector{Float32}}, Array{Float32, 4}}

Extract values of given vector field (either nodal values of a piecewise linear vector field or a callback function providing evaluation of the vector field for given generalized barycentric coordinates). at all sampling points on offset+ i*spacing for i in Z^d defined by the tuples offset and spacing.

By default, offset is at the minimum of grid coordinates, and spacing is defined the largest grid extend divided by 10.

The intermediate rasterflux in future versions can be used to calculate streamlines.

The code is 3D ready.

GridVisualizeTools.extract_visible_bfaces3D — Method

extract_visible_bfaces3D(grid, xyzcut; primepoints, Tp, Tf)

Extract visible boundary faces - those not cut off by the planes x=xyzcut[1] or y=xyzcut[2] or z=xyzcut[3].

Return corresponding points and facets for each region for drawing as mesh (Makie,MeshCat) or trisurf (pyplot)

GridVisualizeTools.extract_visible_cells3D — Method

extract_visible_cells3D(grid, xyzcut; primepoints, Tp, Tf)

Extract visible tetrahedra - those intersecting with the planes x=xyzcut[1] or y=xyzcut[2] or z=xyzcut[3].

Return corresponding points and facets for each region for drawing as mesh (Makie,MeshCat) or trisurf (pyplot)

GridVisualizeTools.marching_tetrahedra — Method

marching_tetrahedra(grid, func, planes, flevels; kwargs...)

Extract isosurfaces and plane interpolation for function on 3D tetrahedral mesh. See marching_tetrahedra(coord,cellnodes,func,planes,flevels;tol, primepoints, primevalues, Tv, Tp, Tf)

GridVisualizeTools.marching_triangles — Method

marching_triangles(grid, func, levels)

Collect isoline snippets on triangles ready for linesegments!

PyPlot

GridVisualize.tridata — Method

tridata(grid)

Return tridata to be splatted to PyPlot calls

Makie

GridVisualize.makeaxis3d — Method

makeaxis3d(ctx)

Dispatch between LScene and new Axis3. Axis3 does not allow zoom, so we support LScene in addition.

GridVisualize.makescene2d — Method

makescene2d(ctx)

Complete scene with title and status line showing interaction state. This uses a gridlayout and its protrusion capabilities.

GridVisualize.makescene3d — Method

   makescene3d(ctx)

Complete scene with title and status line showing interaction state. This uses a gridlayout and its protrusion capabilities.

GridVisualize.scene_interaction — Function

 scene_interaction(update_scene,view,switchkeys::Vector{Symbol}=[:nothing])

Control multiple scene elements via keyboard up/down keys. Each switchkey is assumed to correspond to one of these elements. Pressing a switch key transfers control to its associated element.

Control of values of the current associated element is performed by triggering change values via up/down (± 1) resp. pageup/pagedown (±10) keys

The update_scene callback gets passed the change value and the symbol.

GridVisualize.FlippableLayout.FLayout — Type

mutable struct FLayout

Struct describing flippable layout data. We don't type annotate with Makie types as they are unknown at start time.

visible: Visible GridLayout

offscreen: Hidden GridLayot

blocked: Has the layout been blocked by the block key ?

layoutables: Layoutables attached to layout

condition: Condition variable working together with the blocked field.

GridVisualize.FlippableLayout.flayoutscene — Method

flayoutscene(;blocked=false, kwargs...)

Layoutscene with interactive layout and blocking functionality.

The , key switches between focused view showing only one subscene and "gallery view" showing all layoutables at once.

The space key toggles blocking of the execution of the main therad when yield is replaced by yieldwait. Initial blocking state is set by the blocked kwarg.

The kwargs... are the same as of AbstractPlotting.layoutscene.

The idea is that this can work in some cases as a drop-in replacement of layoutscene.

GridVisualize.FlippableLayout.setmakie! — Method

setmakie!(MyMakie)

Set the Makie module. This Makie can be GLMakie,WGLMakie,CairoMakie

GridVisualize.FlippableLayout.yieldwait — Method

 yieldwait(fliplayoutscene)

Yield and wait in case of scene being blocked via space key toggle

Plots

GridVisualize.rectdata — Method

rectdata(grid, U)

Return rectangular grid data + function to be splatted into Plots calls

VTKView