Private API

Common methods

Base.getindex — Method

getindex(p, i, j)

Return a SubVisualizer

Base.size — Method

size(p)

Return the layout of a GridVisualizer

GridVisualize.ismeshcat — Method

ismeshcat(Plotter)

Heuristically check if Plotter is MeshCat

ColorTypes.RGB — Method

RGB(c)

Create RGB color from color name string.

ColorTypes.RGB — Method

RGB(c)

Create RGB color from color name symbol.

ColorTypes.RGB — Method

RGB(c)

Create RGB color from tuple

GridVisualize.bregion_cmap — Method

bregion_cmap(n)

Create customized distinguishable colormap for boundary regions.

These use fully saturated colors.

GridVisualize.extract_visible_bfaces3D — Method

extract_visible_bfaces3D(coord, bfacenodes, bfaceregions, nbregions, xyzcut; primepoints, Tp, Tf)

Extract visible boundary faces - those not cut off by the planes x=xyzcut[1] or y=xyzcut[2] or z=xyzcut[3].

Return corresponding points and facets for each region for drawing as mesh (Makie,MeshCat) or trisurf (pyplot)

GridVisualize.extract_visible_bfaces3D — Method

extract_visible_bfaces3D(grid, xyzcut; primepoints, Tp, Tf)

Extract visible boundary faces - those not cut off by the planes x=xyzcut[1] or y=xyzcut[2] or z=xyzcut[3].

Return corresponding points and facets for each region for drawing as mesh (Makie,MeshCat) or trisurf (pyplot)

GridVisualize.extract_visible_cells3D — Method

extract_visible_cells3D(coord, cellnodes, cellregions, nregions, xyzcut; primepoints, Tp, Tf)

Extract visible tetrahedra - those intersecting with the planes x=xyzcut[1] or y=xyzcut[2] or z=xyzcut[3].

Return corresponding points and facets for each region for drawing as mesh (Makie,MeshCat) or trisurf (pyplot)

GridVisualize.extract_visible_cells3D — Method

extract_visible_cells3D(grid, xyzcut; primepoints, Tp, Tf)

Extract visible tetrahedra - those intersecting with the planes x=xyzcut[1] or y=xyzcut[2] or z=xyzcut[3].

Return corresponding points and facets for each region for drawing as mesh (Makie,MeshCat) or trisurf (pyplot)

GridVisualize.marching_tetrahedra — Method

marching_tetrahedra(coord, cellnodes, func, planes, flevels; tol, primepoints, primevalues, Tv, Tp, Tf)

Extract isosurfaces and plane interpolation for function on 3D tetrahedral mesh.

The basic observation is that locally on a tetrahedron, cuts with planes and isosurfaces of P1 functions look the same. This method calculates data for several plane cuts and several isosurfaces at once.

Input parameters:

coord: 3 x n_points matrix of point coordinates

cellnodes: 4 x n_cells matrix of point numbers per tetrahedron

func: n_points vector of piecewise linear function values

planes: vector of plane equations ax+by+cz+d=0,each stored as vector [a,b,c,d]

flevels: vector of function isolevels

Keyword arguments:

tol: tolerance for tet x plane intersection

primepoints: 3 x n_prime matrix of "corner points" of domain to be plotted. These are not in the mesh but are used to calculate the axis size e.g. by Makie

primevalues: n_prime vector of function values in corner points. These can be used to calculate function limits e.g. by Makie

Tv: type of function values returned

Tp: type of points returned

Tf: type of facets returned

Return values: (points, tris, values)

points: vector of points (Tp)

tris: vector of triangles (Tf)

values: vector of function values (Tv)

These can be readily turned into a mesh with function values on it.

GridVisualize.marching_tetrahedra — Method

marching_tetrahedra(grid, func, planes, flevels; kwargs...)

Extract isosurfaces and plane interpolation for function on 3D tetrahedral mesh. See marching_tetrahedra(coord,cellnodes,func,planes,flevels;tol, primepoints, primevalues, Tv, Tp, Tf)

GridVisualize.marching_triangles — Method

marching_triangles(coord, cellnodes, func, levels)

Collect isoline snippets on triangles ready for linesegments!

GridVisualize.marching_triangles — Method

marching_triangles(grid, func, levels)

Collect isoline snippets on triangles ready for linesegments!

GridVisualize.markerpoints — Method

markerpoints(points, nmarkers, transform)

Assume that points are nodes of a polyline. Place nmarkers equidistant markers at the polyline, under the assumption that the points are transformed via the transformation matrix M vor visualization.

GridVisualize.quiverdata — Method

quiverdata(rastercoord, rasterflux; vscale, vnormalize) -> Tuple{Union{Base.ReshapedArray{Float32, 2, _A, _B} where {_A<:AbstractArray, _B<:Tuple{Vararg{Base.MultiplicativeInverses.SignedMultiplicativeInverse{Int64}, N} where N}}, Matrix{Float32}}, Union{Base.ReshapedArray{Float32, 2, _A, _B} where {_A<:AbstractArray, _B<:Tuple{Vararg{Base.MultiplicativeInverses.SignedMultiplicativeInverse{Int64}, N} where N}}, Matrix{Float32}}}

Extract nonzero fluxes for quiver plots from rastergrid.

Returns qc, qv - d x nquiver matrices.

If vnormalize is true, the vector field is normalized to vscale*min(spacing), otherwise, it is scaled by vscale Result data are meant to be ready for being passed to calls to quiver.

GridVisualize.region_cmap — Method

region_cmap(n)

Create customized distinguishable colormap for interior regions. For this we use a kind of pastel colors.

GridVisualize.rgbtuple — Method

rgbtuple(c)

Create color tuple from color description (e.g. string)

GridVisualize.rgbtuple — Method

rgbtuple(c)

Create color tuple from RGB color.

GridVisualize.tet_x_plane! — Method

tet_x_plane!(ixcoord, ixvalues, pointlist, node_indices, planeq_values, function_values; tol)

Calculate intersections between tetrahedron with given piecewise linear function data and plane

Adapted from https://github.com/j-fu/gltools/blob/master/glm-3d.c#L341

A non-empty intersection is either a triangle or a planar quadrilateral, define by either 3 or 4 intersection points between tetrahedron edges and the plane.

Input:

  pointlist: 3xN array of grid point coordinates

nodeindices: 4 element array of node indices (pointing into pointlist and functionvalues)

planeq_values: 4 element array of plane equation evaluated at the node coordinates

function_values: N element array of function values

Mutates:

ixcoord: 3x4 array of plane - tetedge intersection coordinates

ixvalues: 4 element array of fuction values at plane - tetdedge intersections

Returns:

nxs,ixcoord,ixvalues

This method can be used both for the evaluation of plane sections and for the evaluation of function isosurfaces.

GridVisualize.vectorsample — Method

vectorsample(grid::ExtendableGrid{Tv, Ti}, v; offset, spacing, reltol) -> Tuple{Vector{Vector{Float32}}, Array{Float32, 4}}

Extract values of given vector field (either nodal values of a piecewise linear vector field or a callback function providing evaluation of the vector field for given generalized barycentric coordinates). at all sampling points on offset+ i*spacing for i in Z^d defined by the tuples offset and spacing.

By default, offset is at the minimum of grid coordinates, and spacing is defined the largest grid extend divided by 10.

The intermediate rasterflux in future versions can be used to calculate streamlines.

The code is 3D ready.

PyPlot

GridVisualize.tridata — Method

tridata(grid)

Return tridata to be splatted to PyPlot calls

Makie

GridVisualize.makeaxis3d — Method

makeaxis3d(ctx)

Dispatch between LScene and new Axis3. Axis3 does not allow zoom, so we support LScene in addition.

GridVisualize.makescene2d — Method

makescene2d(ctx)

Complete scene with title and status line showing interaction state. This uses a gridlayout and its protrusion capabilities.

GridVisualize.scene_interaction — Function

 scene_interaction(update_scene,view,switchkeys::Vector{Symbol}=[:nothing])

Control multiple scene elements via keyboard up/down keys. Each switchkey is assumed to correspond to one of these elements. Pressing a switch key transfers control to its associated element.

Control of values of the current associated element is performed by triggering change values via up/down (± 1) resp. pageup/pagedown (±10) keys

The update_scene callback gets passed the change value and the symbol.

GridVisualize.FlippableLayout.FLayout — Type

mutable struct FLayout

Struct describing flippable layout data. We don't type annotate with Makie types as they are unknown at start time.

visible

Visible GridLayout

offscreen

Hidden GridLayot

blocked

Has the layout been blocked by the block key ?

layoutables

Layoutables attached to layout

condition

Condition variable working together with the blocked field.

GridVisualize.FlippableLayout.flayoutscene — Method

flayoutscene(;blocked=false, kwargs...)

Layoutscene with interactive layout and blocking functionality.

The , key switches between focused view showing only one subscene and "gallery view" showing all layoutables at once.

The space key toggles blocking of the execution of the main therad when yield is replaced by yieldwait. Initial blocking state is set by the blocked kwarg.

The kwargs... are the same as of AbstractPlotting.layoutscene.

The idea is that this can work in some cases as a drop-in replacement of layoutscene.

GridVisualize.FlippableLayout.setmakie! — Method

setmakie!(MyMakie)

Set the Makie module. This Makie can be GLMakie,WGLMakie,CairoMakie

GridVisualize.FlippableLayout.yieldwait — Method

 yieldwait(fliplayoutscene)

Yield and wait in case of scene being blocked via space key toggle

Plots

GridVisualize.rectdata — Method

rectdata(grid, U)

Return rectangular grid data + function to be splatted into Plots calls

VTKView